User Registeration, Delete Room, Add Messages, Participants ve Recent Activities oluşturuldu. CRUD ogeleri oluşturuldul ve database implentasyonu gerçekleşti.

+++ base/views.py
@@ -1,10 +1,15 @@
 
-from multiprocessing import context
+
+from email import message
 from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.contrib.auth.models import User
-from .models import Room, Topic
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Message, Room, Topic
 from .forms import RoomForm
 
 
@@ -28,19 +33,54 @@ from .forms import RoomForm
 
 def loginPage(request):
 
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
             user = User.objects.get(username=username)
         except:
+            messages.error(request, 'Böyle bir Kullanıcı Adı yok')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+
+        else:
             messages.error(request, 'Kullanıcı Adınız veya Şifreniz Hatalı')
 
 
-    context = {}
+
+    context = {'page': page}
     return render(request,'base/login_register.html', context)
 
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+
+        else:
+            messages.error(request, 'Kayıt esnasında bir hata oluştu.')
+
+    return render(request, 'base/login_register.html', {'form':form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -53,18 +93,31 @@ def home(request):
     topics = Topic.objects.all()
 
     room_count = rooms.count()
+    room_messages = Message.objects.all()
 
-    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count, 'room_messages':room_messages}
     return render(request, 'base/home.html', context)
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    
-    context = {'room' : room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
+    context = {'room' : room, 'room_messages': room_messages, 'participants':participants}
 
     return render(request, 'base/room.html', context) 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm
     if request.method == 'POST':
@@ -77,10 +130,13 @@ def createRoom(request):
     return render(request,'base/room_form.html', context)
 
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
      room = Room.objects.get(id=pk)
      form = RoomForm(instance=room)
+
+     if request.user != room.host:
+        return HttpResponse('Bunu yapmak için yetkili değilsiniz.')
 
      if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -91,11 +147,27 @@ def updateRoom(request, pk):
      context =  {'form' : form}
      return render(request, 'base/room_form.html', context)
      
-
+@login_required(login_url='login')
 def deleteRoom(request, pk ):
     room = Room.objects.get(id=pk) 
+
+    if request.user != room.host:
+        return HttpResponse('Bunu yapmak için yetkili değilsiniz.')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj' : room})
     
+
+@login_required(login_url='login')
+def deleteMessage(request, pk ):
+    message = Message.objects.get(id=pk) 
+
+    if request.user != message.user:
+        return HttpResponse('Bunu yapmak için yetkili değilsiniz.')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj' : message})
